requirements.txt
streamlit
faiss-cpu
sentence-transformers
numpy
pickle5
google-generativeai
python-dotenv
